requirements.txt
streamlit==1.2.0
Pillow==10.0.0
